chore/fixed the session cookie issue
