+++ src/actions/session.ts
@@ -1,21 +1,16 @@
+"use server";
+
 import { SessionModel } from "@/models/session";
-import { CookieKeys, getCookie } from "./cookies";
-import { GET } from "@/app/api/set-session/route";
-import { setCookie } from "@/util/Cookies";
+import { CookieKeys, getCookie, setCookie } from "./cookies";
+import { cookies } from "next/headers";
 
-export const ensureSessionID = async () => {
-  const token = getCookie(CookieKeys.SESSION);
-  if (!token) setSession;
-};
-
-const setSession = async () => {
-  const response = await fetch("/api/set-session");
+export const setSession = async () => {
+  const response = await fetch(`${process.env.FED_ROOT}/api/session/create`);
 
   if (!response.ok) {
     throw new Error("failed to get set data");
   }
 
   const data: SessionModel = await response.json();
-
-  setCookie(CookieKeys.SESSION, data.id);
+  return data;
 };
