+++ src/models/cookies.ts
@@ -1,0 +1,3 @@
+export enum CookieKeys {
+  SESSION = "_session",
+}
