+++ src/util/Cookies.ts
@@ -1,22 +1,12 @@
-import { cookies } from "next/headers";
+"use client";
 
-export enum CookieKeys {
-  SESSION = "_session",
-}
-
-export function hasCookie(key: CookieKeys) {
-  return cookies().has(key)
-}
+import { CookieKeys } from "@/models/cookies";
 
 export function getCookie(key: CookieKeys) {
-  return cookies().get(key)
+  var match = document.cookie.match(new RegExp("(^| )" + key + "=([^;]+)"));
+  if (match) return match[2];
 }
 
-export function setCookie(key: CookieKeys, value: string, path?: string, expires?: boolean, httpOnly?: boolean) {
-  cookies().set({
-    name: key,
-    value: value,
-    path: path ?? '/',
-    httpOnly: httpOnly ?? true 
-  })
-};
+export function setCookie(key: CookieKeys, value: string) {
+  document.cookie = `${key}=${value}`;
+}
